Support eBay's new item conditions
eBay has established a new way of reproting the condition of an auction.
Each listing category has it's own set of possible conditions.  A
condition consists of a canonical numeric identifier along with a
natural language description.

The new condition method on the Item class returns the information if
it's available.

Closes gh-1